Fixes the formatting of the class time
Since starting_at and finishes_at are Ecto.Time format it cannot be directly formatted by Timex.

Timex first need to convert them to naive_datetime and that need a full DateTime.

To solve this issue we need to use the class date to convert class times to DateTime and then format them as we wish